charts/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
Django==3.0.8
django-bootstrap4==2.2.0
django-chartjs==2.2.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
